split ERC721 tests into behaviour
